Класс Node перенесен в InMemoryHistoryManager
